+++ trunk/symmetric/symmetric-core/src/main/java/org/jumpmind/symmetric/service/impl/TriggerRouterService.java
@@ -587,7 +587,7 @@ public class TriggerRouterService extends AbstractService implements ITriggerRou
                         triggerRouter.getLastUpdateTime(),
                         triggerRouter.getTrigger().getTriggerId(),
                         triggerRouter.getRouter().getRouterId() },
-                new int[] { Types.INTEGER, Types.VARCHAR, Types.SMALLINT, Types.VARCHAR, Types.TIMESTAMP,
+                new int[] { Types.NUMERIC, Types.VARCHAR, Types.SMALLINT, Types.VARCHAR, Types.TIMESTAMP,
                         Types.VARCHAR, Types.VARCHAR })) {
             triggerRouter.setCreateTime(triggerRouter.getLastUpdateTime());
             jdbcTemplate.update(getSql("insertTriggerRouterSql"), new Object[] {
@@ -595,7 +595,7 @@ public class TriggerRouterService extends AbstractService implements ITriggerRou
                     triggerRouter.isPingBackEnabled() ? 1 : 0,
                     triggerRouter.getCreateTime(), triggerRouter.getLastUpdateBy(),
                     triggerRouter.getLastUpdateTime(), triggerRouter.getTrigger().getTriggerId(),
-                    triggerRouter.getRouter().getRouterId() }, new int[] { Types.INTEGER,
+                    triggerRouter.getRouter().getRouterId() }, new int[] { Types.NUMERIC,
                     Types.VARCHAR, Types.SMALLINT, Types.TIMESTAMP, Types.VARCHAR, Types.TIMESTAMP, Types.VARCHAR,
                     Types.VARCHAR });
         }
